+++ src/main/kotlin/SqlCipherPlugin.kt
@@ -60,8 +60,9 @@ class SqlCipherPlugin: Plugin<Project> {
         val ext = SqlcipherExtension.createExtensions(target)
         val opensslTask = OpenSslBuild(target, ext)
         val sqlcipherTask = SqlCipherBuild(target, ext)
-        registerAll(target, opensslTask.buildName, ext.builds, false)
-        registerAll(target, sqlcipherTask.buildName, ext.builds, true)
+        val allBuilds = BuildType.values().toList()
+        registerAll(target, opensslTask.buildName, allBuilds, false)
+        registerAll(target, sqlcipherTask.buildName, allBuilds, true)
 
         target.tasks.register("${opensslTask.buildName}Clean", CleanAllTask::class.java)
             .configure{